🔨 [projects] Move constants `LATEST_BRANCH` and `UPDATE_BRANCH` to `repository`
