+++ tests/unit/projects/test_create.py
@@ -8,8 +8,8 @@ from cutty.filestorage.adapters.disk import DiskFileStorage
 from cutty.filestorage.domain.files import RegularFile
 from cutty.filestorage.domain.storage import FileStorage
 from cutty.filesystems.domain.purepath import PurePath
-from cutty.projects.common import LATEST_BRANCH
 from cutty.projects.loadtemplate import TemplateMetadata
+from cutty.projects.repository import LATEST_BRANCH
 from cutty.projects.repository import ProjectRepository
 from cutty.util.git import Repository
